# Feat/batched gpu inference clean