… .gitattributes
… .github/dependabot.yml
… .github/workflows/ci.yml
… .github/workflows/mmdet.yml
… .github/workflows/publish_docs.yml
… .github/workflows/publish_pypi.yml
… .github/workflows/ruff.yaml
… .gitignore
… CHANGELOG.md
… CITATION.cff
… CONTRIBUTING.md
… LICENSE
… MANIFEST.in
… README.md
… context7.json
… demo/demo_data/terrain2_coco.json
… demo/inference_for_detectron2.ipynb
… demo/inference_for_huggingface.ipynb
… demo/inference_for_mmdetection.ipynb
… demo/inference_for_roboflow.ipynb
… demo/inference_for_rtdetr.ipynb
… demo/inference_for_torchvision.ipynb
… demo/inference_for_yolov5.ipynb
… demo/slicing.ipynb
… docs/annotation.md
… docs/auto_model.md
… docs/changelog.md
… docs/cli.md
… docs/coco.md
… docs/contributing.md